requirements.txt
numpy>=1.24
scipy>=1.10
soundfile>=0.12
pyyaml>=6
rich>=13
scikit-learn>=1.3
hdbscan>=0.8
musicnn>=0.2
mutagen>=1.46
PyDejavu>=0.6
librosa>=0.10
faiss-cpu>=1.7
laion-clap @ git+https://github.com/LAION-AI/CLAP.git#egg=laion_clap